requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.4
streamlit
opencv-python
numpy
Pillow
streamlit
opencv-python
numpy
Pillow
streamlit-image-comparison
streamlit-drawable-canvas
streamlit-ace
streamlit-option-menu
streamlit-lottie
streamlit-plotly-events
streamlit-graphviz
streamlit-aggrid
streamlit-extras
streamlit-ace
streamlit-plotly-events
streamlit-graphviz
streamlit-aggrid
streamlit-extras
streamlit-plotly-events
streamlit-graphviz
